debian/patches/08-test.sh-fixes.patch: exclude failing tests
We do not use socat in our product. It accessible on the box and nothing
uses that. Previously rabbitmq depends on that between the rabbitmq
version 3.6 and 3.9. From version 3.9 rabbitmq drops the needing of
socat. You can read it from the change log here:
https://git.balabit/balabit-os-9.0/rabbitmq-server/-/blob/master/release-notes/3.9.0.md
In opposite Ubuntu has not removed this dependency from the
rabbitmq-server package yet.
In the past we try to resolve the failing tests with no success:
https://git.balabit/balabit-os-8.0/socat/-/commit/6f5e9eca9b30d2e07c092c04372cb1be26ac2db0
Because of nothing and nobody want to use socat and probably is not used
in the future, we just exclude the failing tests.